Re-arrange and remove old code
